% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Compute confusion matrix}
\usage{
confusion_matrix(true, pred)
}
\arguments{
\item{true}{true categories}

\item{pred}{predicted categories (last column of the data.frame output by \code{predict.gbdt})}
}
\description{
Compute confusion matrix
}

