% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confusion_proba_plot}
\alias{confusion_proba_plot}
\title{Plot the distribution of signals in categories after prediction}
\usage{
confusion_proba_plot(true, pred, prob)
}
\arguments{
\item{true}{true categories}

\item{pred}{predicted categories (last column of the data.frame output by \code{predict.gbdt})}

\item{prob}{for each signal, probabilities to be in each category}

\item{...}{passed to \code{geom_??}}
}
\description{
Plot the distribution of signals in categories after prediction
}
\examples{
data(sirena)
sub <- subsample(sirena[,-ncol(sirena)], p=0.2)
train <- sirena[sub$picked,]
data <- sirena[!sub$picked,]
p <- classify(data=data[,-ncol(data)], train=train,
              n.trees=100, shrinkage=0.01, n.minobsinnode=1,
              n.cores=1)
confusion_proba_plot(true=data$type, pred=p$type, prob=p[,1:4])
}
\seealso{
\code{gbm}, \code{summary.gbm}, \code{predict.gbm} in package \code{gbm}
}

