% Generated by roxygen2 (4.0.1): do not edit by hand
\name{confusion_stats}
\alias{confusion_stats}
\title{Compute confusion statistics (recall, precision, etc.)}
\usage{
confusion_stats(x)
}
\arguments{
\item{x}{a confusion matrix (table with predictions as lines and true observations as columns, as output by \code{confusion_matrix})}
}
\description{
Compute confusion statistics (recall, precision, etc.)
}

