% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.subsample}
\alias{plot.subsample}
\title{Plot subsampled partition}
\usage{
\method{plot}{subsample}(x, ...)
}
\arguments{
\item{x}{data.frame output by the \code{subsample()} function}

\item{...}{passed to and from other methods}
}
\description{
Plot subsampled partition
}
\seealso{
\code{\link{subsample}}
}

