% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classify}
\alias{classify}
\title{Fit GBDT model and predict classification}
\usage{
classify(data, train, verbose = TRUE, ...)
}
\arguments{
\item{data}{a data.frame with signal characteristics of signals to be predicted}

\item{train}{a data.frame with signal characteristics and classification of signals from which to learn the model}

\item{verbose}{when TRUE (the default), print information about the fitted model}

\item{...}{passed to \code{fit.gbdt}}
}
\description{
Fit GBDT model and predict classification
}
\examples{
data(sirena)
sub <- subsample(sirena[,-ncol(sirena)], p=0.2)
train <- sirena[sub$picked,]
data <- sirena[!sub$picked,]
pred <- classify(data=data[,-ncol(data)], train=train, n.trees=100, shrinkage=0.01, n.minobsinnode=1)
head(pred)
(cm <- confusion_matrix(true=data$type, pred=pred$type))
confusion_stats(cm)
}

