% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classify_file}
\alias{classify_file}
\title{Fit GBDT model and predict classification}
\usage{
classify_file(data, train, ...)
}
\arguments{
\item{data}{a tab or space delimited file with signal characteristics of signals to be predicted}

\item{train}{a tab or space delimited file with signal characteristics and classification of signals from which to learn the model}

\item{...}{passed to \code{fit.gbdt}}
}
\value{
A file with the original data and the predictions (probabilities and predicted type)
}
\description{
Fit GBDT model and predict classification
}
\seealso{
\code{\link{classify}} for an example of usage
}

