% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit.gbdt}
\alias{fit.gbdt}
\title{Fit a Gradient Boosting Decision Tree model}
\usage{
fit.gbdt(x, n.trees = 5000, shrinkage = 0.001, interaction.depth = 3, ...)
}
\arguments{
\item{x}{a data.frame with signal characteristics in the first columns and signal type (classification) in the last column}

\item{n.trees}{see \code{?gbm}}

\item{shrinkage}{see \code{?gbm}}

\item{interaction.depth}{see \code{?gbm}}

\item{...}{passed to \code{gbm}}
}
\description{
Fit a Gradient Boosting Decision Tree model
}
\examples{
data(sirena)
head(sirena)
m <- fit.gbdt(x=sirena, n.trees=3, shrinkage=0.01, interaction.depth=1)
# NB: the parameters are caricatural here
#     for the example to run quickly enough and because the data is easy
print(m)
summary(m)
pred <- predict(m, newdata=sirena)
head(pred)
table(true=sirena$type, predicted=pred$type)
}
\seealso{
\code{gbm}, \code{summary.gbm}, \code{predict.gbm} in package \code{gbm}
}

