% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subsample_file}
\alias{subsample_file}
\title{Feature-based subsampling of signals from a file}
\usage{
subsample_file(file, p, k = 10, plot = TRUE)
}
\arguments{
\item{file}{path to a tab or space delimited file containing signal characteristics}

\item{plot}{wether to plot the result of the subsampling}

\item{p}{proportion of the total number of signals to subsample, between 0 and 1}

\item{k}{number of groups}
}
\value{
Two files, one for the signals picked in the subsample and one for the non-picked signals
}
\description{
Read a tab or space delimited file containing signal characteristics and subsample a given proportion of signals
}

