% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gbdt}
\alias{summary.gbdt}
\title{Compute the relative influence of each variable in the model}
\usage{
\method{summary}{gbdt}(object, n.trees = object$best.iter, ...)
}
\arguments{
\item{object}{a gbdt object, from fit.gbdt}

\item{n.trees}{number of trees used in the computation.}

\item{...}{passed to \code{summary.gbm}}
}
\description{
Compute the relative influence of each variable in the model
}
\examples{
data(sirena)
m <- fit.gbdt(x=sirena, n.trees=3, shrinkage=0.01, interaction.depth=1,
              n.cores=1)
# NB: The arguments values are caricatural here for the example to run
#     quickly enough and because the data is easy.
# NB: n.cores = 1 is necessary for examples to run on all machines. Feel
#     free to remove it and use more cores on your machine.
summary(m)
}
\seealso{
\code{summary.gbm} in package \code{gbm}
}

